SWC-6896 - Additional refactoring to consolidate table rendering and separate business logic
