+++ packages/synapse-react-client/src/components/Forum/ForumTable.tsx
@@ -11,11 +11,9 @@ import IconSvg from '../IconSvg/IconSvg'
 import UserCard from '../UserCard/UserCard'
 import { Button, Link } from '@mui/material'
 import { UserBadge } from '../UserCard/UserBadge'
-import { StyledTableContainer } from '../styled/StyledTableContainer'
 import {
   ColumnDef,
   createColumnHelper,
-  flexRender,
   getCoreRowModel,
   SortingState,
   Table,
@@ -23,6 +21,7 @@ import {
 } from '@tanstack/react-table'
 import ColumnHeader from '../styled/ColumnHeader'
 import { isEmpty } from 'lodash-es'
+import StyledTanstackTable from '../Table/Table'
 
 export type ForumTableProps = {
   forumId: string
@@ -153,70 +152,10 @@ export const ForumTable: React.FC<ForumTableProps> = ({
 
   return (
     <div className="ForumTable">
-      <StyledTableContainer
-        sx={{
-          my: 2,
-          ['th,td']: { px: 1 },
-          td: {
-            py: 1,
-          },
-        }}
-      >
-        <table style={{ borderCollapse: 'collapse', width: '100%' }}>
-          <thead>
-            {table.getHeaderGroups().map(headerGroup => {
-              return (
-                <tr key={headerGroup.id}>
-                  {headerGroup.headers.map(header => (
-                    <th
-                      key={header.id}
-                      colSpan={header.colSpan}
-                      style={{ width: header.getSize() }}
-                    >
-                      {header.isPlaceholder
-                        ? null
-                        : flexRender(
-                            header.column.columnDef.header,
-                            header.getContext(),
-                          )}
-                      {header.column.getCanResize() && (
-                        <div
-                          className={`resizer ${
-                            header.column.getIsResizing() ? 'isResizing' : ''
-                          }`}
-                          onMouseDown={header.getResizeHandler()}
-                          onTouchStart={header.getResizeHandler()}
-                        />
-                      )}
-                    </th>
-                  ))}
-                </tr>
-              )
-            })}
-          </thead>
-          <tbody>
-            {table.getRowModel().rows.map(row => (
-              <tr key={row.id}>
-                {row.getVisibleCells().map(cell => {
-                  return (
-                    <td
-                      key={cell.id}
-                      style={{
-                        width: cell.column.getSize(),
-                      }}
-                    >
-                      {flexRender(
-                        cell.column.columnDef.cell,
-                        cell.getContext(),
-                      )}
-                    </td>
-                  )
-                })}
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </StyledTableContainer>
+      <StyledTanstackTable
+        table={table}
+        styledTableContainerProps={{ sx: { my: 2 } }}
+      />
       {hasNextPage && (
         <Button
           variant="outlined"
