+++ packages/synapse-react-client/src/components/OAuthClientManagement/OAuthManagement.tsx
@@ -12,16 +12,15 @@ import { displayToast } from '../ToastMessage'
 import { DialogBase } from '../DialogBase'
 import { Box, Button, Link } from '@mui/material'
 import { AddCircleTwoTone } from '@mui/icons-material'
-import { StyledTableContainer } from '../styled/StyledTableContainer'
 import {
   ColumnDef,
   createColumnHelper,
-  flexRender,
   getCoreRowModel,
   Table,
   useReactTable,
 } from '@tanstack/react-table'
 import ColumnHeader from '../styled/ColumnHeader'
+import StyledTanstackTable from '../Table/Table'
 
 const columnHelper = createColumnHelper<OAuthClient>()
 
@@ -156,68 +155,7 @@ export const OAuthManagement: React.FunctionComponent = () => {
           Create New Client
         </Button>
       </Box>
-      <StyledTableContainer
-        sx={{
-          td: {
-            py: 1,
-          },
-        }}
-      >
-        <table style={{ borderCollapse: 'collapse', width: '100%' }}>
-          <thead>
-            {table.getHeaderGroups().map(headerGroup => {
-              return (
-                <tr key={headerGroup.id}>
-                  {headerGroup.headers.map(header => (
-                    <th
-                      key={header.id}
-                      colSpan={header.colSpan}
-                      style={{ width: header.getSize() }}
-                    >
-                      {header.isPlaceholder
-                        ? null
-                        : flexRender(
-                            header.column.columnDef.header,
-                            header.getContext(),
-                          )}
-                      {header.column.getCanResize() && (
-                        <div
-                          className={`resizer ${
-                            header.column.getIsResizing() ? 'isResizing' : ''
-                          }`}
-                          onMouseDown={header.getResizeHandler()}
-                          onTouchStart={header.getResizeHandler()}
-                        />
-                      )}
-                    </th>
-                  ))}
-                </tr>
-              )
-            })}
-          </thead>
-          <tbody>
-            {table.getRowModel().rows.map(row => (
-              <tr key={row.id}>
-                {row.getVisibleCells().map(cell => {
-                  return (
-                    <td
-                      key={cell.id}
-                      style={{
-                        width: cell.column.getSize(),
-                      }}
-                    >
-                      {flexRender(
-                        cell.column.columnDef.cell,
-                        cell.getContext(),
-                      )}
-                    </td>
-                  )
-                })}
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </StyledTableContainer>
+      <StyledTanstackTable table={table} />
       {hasNextPage && (
         <div className="text-center">
           <Button
